Improved performance a bit
key change here is I identified the source of the lag between submitting the form and getting a progress bar up. Related slow celery inspection calls that are not necessary at all. So for now subthreaded them, and rearranged a little. 

Another key change is I incorporated all results the get_updates returns) into the pulse checker loop now. For which reason we needed to add a seen flag to the update queue payloads.